+++ src/components/singleButton.js
@@ -1,6 +1,6 @@
 const { ButtonBuilder, ButtonStyle } = require('discord.js');
 
-const buildSingleButton = (label, data, primary, disabled, eventName) => {
+const buildSingleButton = (label, emoji, data, primary, disabled, eventName) => {
     const buttonId = {
         eventName: eventName,
         ...data,
@@ -8,9 +8,14 @@ const buildSingleButton = (label, data, primary, disabled, eventName) => {
     
     const button = new ButtonBuilder()
         .setCustomId(`${JSON.stringify(buttonId)}`)
-        .setLabel(label)
         .setStyle(primary ? ButtonStyle.Primary : ButtonStyle.Secondary)
         .setDisabled(disabled);
+    if (label) {
+        button.setLabel(label);
+    }
+    if (emoji) {
+        button.setEmoji(emoji);
+    }
     return button;
 }