+++ src/config/gachaConfig.js
@@ -1,5 +1,7 @@
+const seedrandom = require("seedrandom");
 const { backpackItems } = require('./backpackConfig');
-const { rarities } = require('./pokemonConfig');
+const { rarities, rarityBins } = require('./pokemonConfig');
+const { drawIterable } = require("../utils/gachaUtils");
 
 const NUM_DAILY_REWARDS = 3;
 
@@ -10,32 +12,109 @@ const dailyRewardChances = {
     [backpackItems.MASTERBALL]: 0.01,
 }
 
-const pokeballChanceConfig = {
+const pokeballConfig = {
     [backpackItems.POKEBALL]: {
-        [rarities.COMMON]: 0.7,
-        [rarities.RARE]: 0.25,
-        [rarities.EPIC]: 0.04,
-        [rarities.LEGENDARY]: 0.01,
+        chances: {
+            [rarities.COMMON]: 0.7,
+            [rarities.RARE]: 0.25,
+            [rarities.EPIC]: 0.04,
+            [rarities.LEGENDARY]: 0.01,
+        },
+        pity: 1
     },
     [backpackItems.GREATBALL]: {
-        [rarities.COMMON]: 0.3,
-        [rarities.RARE]: 0.55,
-        [rarities.EPIC]: 0.12,
-        [rarities.LEGENDARY]: 0.03,
+        chances: {
+            [rarities.COMMON]: 0.3,
+            [rarities.RARE]: 0.55,
+            [rarities.EPIC]: 0.12,
+            [rarities.LEGENDARY]: 0.03,
+        },
+        pity: 3
     },
     [backpackItems.ULTRABALL]: {
-        [rarities.RARE]: 0.45,
-        [rarities.EPIC]: 0.5,
-        [rarities.LEGENDARY]: 0.05,
+        chances: {
+            [rarities.RARE]: 0.45,
+            [rarities.EPIC]: 0.5,
+            [rarities.LEGENDARY]: 0.05,
+        },
+        pity: 5
     },
     [backpackItems.MASTERBALL]: {
-        [rarities.EPIC]: 0.9,
-        [rarities.LEGENDARY]: 0.1,
+        chances: {
+            [rarities.EPIC]: 0.9,
+            [rarities.LEGENDARY]: 0.1,
+        },
+        pity: 10
     }
 }
+
+const bannerTypes = {
+    STANDARD: "standard",
+    ROTATING: "rotating",
+    SPECIAL: "special",
+}
+
+const bannerTypeConfig = {
+    [bannerTypes.STANDARD]: {
+        "name": "Standard",
+        "description": "Standard banner with all currently available gacha Pokemon.",
+    },
+    [bannerTypes.ROTATING]: {
+        "name": "Rotating",
+        "description": "Banner which rotates rate-up Pokemon randomly every day.",
+    },
+    [bannerTypes.SPECIAL]: {
+        "name": "Special",
+        "description": "Special banner with a limited-time or hand-chosen rate-up Pokemon.",
+    },
+}
+
+const bannerConfig = [
+    {
+        "bannerType": bannerTypes.STANDARD,
+        "name": "Standard",
+        "description": "Standard banner with all currently available gacha Pokemon.",
+        "rateUp": () => {},
+    },
+    {
+        "bannerType": bannerTypes.ROTATING,
+        "name": "Rotating",
+        "description": "Banner which rotates rate-up Pokemon randomly every day.",
+        "rateUp": () => {
+            // seedrandom using date 
+            const date = new Date().getUTCDate();
+            const rng = seedrandom(date);
+            return {
+                [rarities.LEGENDARY]: drawIterable(rarityBins[rarities.LEGENDARY], 1, {
+                    replacement: false,
+                    rng: rng,
+                }),
+                [rarities.EPIC]: drawIterable(rarityBins[rarities.EPIC], 3, {
+                    replacement: false,
+                    rng: rng,
+                }),
+            }
+        }
+    },
+]
+
+const MAX_PITY = 100;
+
+/*const date = new Date();
+date.setUTCMilliseconds(0);
+date.setUTCSeconds(0);
+date.setUTCMinutes(0);
+date.setUTCHours(0);
+console.log(date.toUTCString());
+console.log(date.getTime() / 86400000);
+console.log(Math.floor(new Date().getTime() / 86400000));*/
 
 module.exports = {
     dailyRewardChances,
     NUM_DAILY_REWARDS,
-    pokeballChanceConfig,
+    pokeballConfig,
+    bannerTypes,
+    bannerTypeConfig,
+    bannerConfig,
+    MAX_PITY,
 };