+++ src/services/gacha.js
@@ -1,16 +1,22 @@
 const { backpackCategories, backpackItemConfig, backpackItems } = require('../config/backpackConfig');
-const { dailyRewardChances, NUM_DAILY_REWARDS, pokeballChanceConfig } = require('../config/gachaConfig');
+const { dailyRewardChances, NUM_DAILY_REWARDS, pokeballConfig, bannerConfig, bannerTypes, MAX_PITY } = require('../config/gachaConfig');
 const { rarityBins, pokemonConfig, rarities, rarityConfig } = require('../config/pokemonConfig');
 const { collectionNames } = require('../config/databaseConfig');
 const { updateDocument, insertDocument, countDocuments, QueryBuilder } = require('../database/mongoHandler');
 const { stageNames } = require('../config/stageConfig');
 const { drawDiscrete, drawIterable, drawUniform } = require('../utils/gachaUtils');
-const { MAX_POKEMON } = require('../config/trainerConfig');
+const { MAX_POKEMON, trainerFields } = require('../config/trainerConfig');
+const { getTrainer } = require('./trainer');
+const { getState } = require('./state');
 
 const { logger } = require('../log');
 const { getOrSetDefault } = require('../utils/utils');
 const { calculatePokemonStats, getBattleEligible } = require('./pokemon');
 const { getPokemonExpNeeded } = require('../utils/pokemonUtils');
+const { buildBannerEmbed } = require('../embeds/pokemonEmbeds');
+const { buildScrollActionRow } = require('../components/scrollActionRow');
+const { eventNames } = require('../config/eventConfig');
+const { buildButtonActionRow } = require('../components/buttonActionRow');
 
 const DAILY_MONEY = 300;
 
@@ -162,11 +168,18 @@ const beginnerRoll = (trainer, quantity) => {
     return rolls;
 }
 
-const usePokeball = async (trainer, pokeballId, quantity=1) => {
+const usePokeball = async (trainer, pokeballId, bannerIndex, quantity=1) => {
     // validate quantity
     if (quantity < 1 || quantity > 10) {
         return { data: null, err: "You may only catch between 1-10 Pokemon at once!" };
     }
+
+    // get banner data
+    const bannerData = bannerConfig[bannerIndex];
+    if (!bannerData) {
+        return { data: null, err: "Invalid banner!" };
+    }
+    const bannerType = bannerData.bannerType;
 
     // check for max pokemon
     try {
@@ -187,21 +200,73 @@ const usePokeball = async (trainer, pokeballId, quantity=1) => {
         return { data: null, err: `Not enough Pokeballs of type ${backpackItemConfig[pokeballId].name}! View your items with \`/backpack\`. Use the \`/pokemart\`, daily rewards, or level rewards to get more.` };
     }
 
-    // if pokeball and beginner rolls < 10, roll beginner rolls
+    // draw rarities
+    const drawnRarities = drawDiscrete(pokeballConfig[pokeballId].chances, quantity);
+
+    // draw pokemon, accounting for pity
+    const trainerBannerInfo = getOrSetDefault(trainer, "banners", trainerFields.banners.default)[bannerType];
+    const pokemonIds = [];
+    const bannerRateUps = bannerData.rateUp() || {};
+    // if pokeball and standard banner and beginner rolls < 10, roll beginner rolls
     const currentNumRolls = getOrSetDefault(trainer, "beginnerRolls", 0);
     let beginnerRolls = null;
-    if (pokeballId == backpackItems.POKEBALL && currentNumRolls < 10) {
+    if (pokeballId == backpackItems.POKEBALL && bannerType === bannerTypes.STANDARD &&  currentNumRolls < 10) {
         beginnerRolls = beginnerRoll(trainer, quantity);
     }
+    // get non-rate-ups for each rarity
+    const nonRateUps = {};
+    for (const rarity in rarityBins) {
+        const rarityRateUps = bannerRateUps[rarity] || [];
+        nonRateUps[rarity] = rarityBins[rarity].filter(pokemonId => !rarityRateUps.includes(pokemonId));
+    }
+    // roll for pokemon
+    for (const rarity of drawnRarities) {
+        let rarityRateUp = bannerRateUps[rarity];
+        if (trainerBannerInfo.pity >= MAX_PITY) {
+            // set rarity to legendary
+            rarityRateUp = bannerRateUps[rarities.LEGENDARY];
+            if (rarityRateUp) {
+                pokemonIds.push(drawIterable(rarityRateUp, 1)[0]);
+            } else {
+                pokemonIds.push(drawIterable(rarityBins[rarities.LEGENDARY], 1)[0]);
+            }
 
-    // draw rarities
-    const rarities = drawDiscrete(pokeballChanceConfig[pokeballId], quantity);
+            // reset pity
+            trainerBannerInfo.pity -= MAX_PITY + 1;
+        } else {
+            // if rate up and 50% chance
+            if (rarityRateUp) {
+                if (Math.random() < 0.5) {
+                    pokemonIds.push(drawIterable(rarityRateUp, 1)[0]);
+                    if (rarity === rarities.LEGENDARY) {
+                        // reset pity
+                        trainerBannerInfo.pity -= MAX_PITY + 1;
+                    }
+                } else {
+                    pokemonIds.push(drawIterable(nonRateUps[rarity], 1)[0]);
+                }
+            } else {
+                pokemonIds.push(drawIterable(nonRateUps[rarity], 1)[0]);
+                if (rarity === rarities.LEGENDARY) {
+                    // reset pity
+                    trainerBannerInfo.pity -= MAX_PITY + 1;
+                }
+            }
+        }
+
+        // add pity based on pokeball pity
+        trainerBannerInfo.pity += pokeballConfig[pokeballId].pity;
+    }
 
     try {
         const res = await updateDocument(
             collectionNames.USERS, 
             { userId: trainer.userId }, 
-            { $set: { backpack: trainer.backpack, beginnerRolls: trainer.beginnerRolls } }
+            { $set: { 
+                backpack: trainer.backpack, 
+                beginnerRolls: trainer.beginnerRolls ,
+                banners: trainer.banners
+            } }
         );
         if (res.modifiedCount === 0) {
             logger.error(`Failed to use pokeball and update ${trainer.user.username}.`);
@@ -213,7 +278,6 @@ const usePokeball = async (trainer, pokeballId, quantity=1) => {
         return { data: null, err: "Error using Pokeball." };
     }
 
-    const pokemonIds = rarities.map(rarity => drawIterable(rarityBins[rarity], 1)[0]);
     if (beginnerRolls) {
         for (const [index, pokemonId] of Object.entries(beginnerRolls)) {
             pokemonIds[index] = pokemonId;
@@ -223,9 +287,145 @@ const usePokeball = async (trainer, pokeballId, quantity=1) => {
     return await giveNewPokemons(trainer, pokemonIds);
 }
 
+const buildBannerSend = async ({ stateId=null, user=null, button=null, page=null } = {}) => {
+    // get state
+    const state = getState(stateId);
+
+    // check page
+    if (page !== null) {
+        if (page < 1 || page > bannerConfig.length) {
+            return { send: null, err: "Invalid banner." };
+        }
+        state.index = page - 1;
+    }
+
+    // if pokeball undefined, default to Pokeball
+    if (state.pokeballId === undefined) {
+        state.pokeballId = backpackItems.POKEBALL;
+    }
+
+    // get trainer
+    let trainer = await getTrainer(user);
+    if (trainer.err) {
+        return { send: null, err: trainer.err };
+    }
+    trainer = trainer.data;
+
+    const send = {
+        embeds: [],
+        components: []
+    }
+
+    // banner embed
+    const bannerData = bannerConfig[state.index];
+    const bannerEmbed = buildBannerEmbed(trainer, bannerData);
+    send.embeds.push(bannerEmbed);
+
+    // scroll buttons
+    const scrollData = {
+        stateId: stateId,
+    }
+    const scrollButtons = buildScrollActionRow(
+        state.index + 1, 
+        state.index === bannerConfig.length - 1,
+        scrollData,
+        eventNames.BANNER_SCROLL
+    );
+    send.components.push(scrollButtons);
+
+    // pokeball select buttons
+    const pokeballData = {
+        stateId: stateId,
+    }
+    const pokeballButtonConfigs = [
+        {
+            data: {
+                ...pokeballData,
+                pokeballId: backpackItems.POKEBALL,
+            },
+            disabled: state.pokeballId === backpackItems.POKEBALL,
+            emoji: backpackItemConfig[backpackItems.POKEBALL].emoji,
+        }, 
+        {
+            data: {
+                ...pokeballData,
+                pokeballId: backpackItems.GREATBALL,
+            },
+            disabled: state.pokeballId === backpackItems.GREATBALL,
+            emoji: backpackItemConfig[backpackItems.GREATBALL].emoji,
+        },
+        {
+            data: {
+                ...pokeballData,
+                pokeballId: backpackItems.ULTRABALL,
+            },
+            disabled: state.pokeballId === backpackItems.ULTRABALL,
+            emoji: backpackItemConfig[backpackItems.ULTRABALL].emoji,
+        },
+        {
+            data: {
+                ...pokeballData,
+                pokeballId: backpackItems.MASTERBALL,
+            },
+            disabled: state.pokeballId === backpackItems.MASTERBALL,
+            emoji: backpackItemConfig[backpackItems.MASTERBALL].emoji,
+        },
+        {
+            label: "Info",
+            data: {
+                ...pokeballData
+            },
+            disabled: false,
+        }
+    ]
+    const pokeballButtons = buildButtonActionRow(
+        pokeballButtonConfigs,
+        eventNames.BANNER_BUTTON
+    )
+    send.components.push(pokeballButtons);
+
+    // build choose quantity select gacha
+    const pokeballs = getOrSetDefault(trainer.backpack, backpackCategories.POKEBALLS, {});
+    const numPokeballs = getOrSetDefault(pokeballs, state.pokeballId, 0);
+    const maxNumRolls = Math.min(numPokeballs, 10);
+    const gachaData = {
+        stateId: stateId,
+    }
+    const gachaButtonConfigs = [
+        {
+            label: "x1",
+            data: {
+                ...gachaData,
+                quantity: 1,
+            },
+            disabled: numPokeballs < 1,
+            emoji: backpackItemConfig[state.pokeballId].emoji,
+        },
+    ]
+    if (maxNumRolls > 1) {
+        gachaButtonConfigs.push({
+            label: `x${maxNumRolls}`,
+            data: {
+                ...gachaData,
+                quantity: maxNumRolls == 1 ? 0 : maxNumRolls,
+            },
+            disabled: numPokeballs < maxNumRolls,
+            emoji: backpackItemConfig[state.pokeballId].emoji,
+        });
+    }
+    const gachaButtons = buildButtonActionRow(
+        gachaButtonConfigs,
+        eventNames.BANNER_GACHA
+    )
+    send.components.push(gachaButtons);
+
+    return { send: send, err: null };
+}
+
 module.exports = {
     drawDaily,
     giveNewPokemons,
     usePokeball,
     generateRandomPokemon,
+    buildBannerSend,
 };