+++ src/utils/gachaUtils.js
@@ -14,10 +14,24 @@ const drawDiscrete = (probabilityDistribution, times) => {
     return results;
 }
 
-const drawIterable = (iterable, times) => {
+const drawIterable = (iterable, times, { replacement=true, rng=Math.random } = {}) => {
+    // check replacement and times
+    if (!replacement && times > iterable.length) {
+        throw new Error("Cannot draw without replacement more times than there are items in the iterable.");
+    }
+
     const results = [];
-    for (let i = 0; i < times; i++) {
-        results.push(iterable[Math.floor(Math.random() * iterable.length)]);
+    if (replacement) {
+        for (let i = 0; i < times; i++) {
+            results.push(iterable[Math.floor(rng() * iterable.length)]);
+        }
+    } else {
+        const iterableCopy = iterable.slice();
+        for (let i = 0; i < times; i++) {
+            const randIndex = Math.floor(rng() * iterableCopy.length);
+            results.push(iterableCopy[randIndex]);
+            iterableCopy.splice(randIndex, 1);
+        }
     }
     return results;
 }